Add specialized augmented/diminished parser
This needs to be seperated from the main chord parser for a number of
reasons.

An augmented chord is only found with a seventh extensions. The upper
extensions are usually not seen, with it being customary to instead
apply the fifth alteration from the alteration list.

A diminished chord with a seventh is often referred to as a
fully-diminished seventh chord. This has an implied `bb7` which can not
be handled in the standard case. The half-diminished case for the moment
is strictly denoted with a `m7b5` but support for the explicit symbol
can be added fairly easily.